refactor : 페이지 내부 폴더 삭제, 필요 없는 이미지 삭제
