+++ src/App.js
@@ -1,7 +1,7 @@
 import './App.css';
 
-import BoothDetailPage from './pages/boothpage/BoothDetailPage';
-import PerfDetailPage from './pages/perfpage/PerfDetailPage';
+import BoothDetailPage from './pages/BoothDetailPage';
+import PerfDetailPage from './pages/PerfDetailPage';
 
 function App() {
   return (
