+++ src/pages/BoothDetailPage.js
@@ -1,11 +1,11 @@
 import styled from 'styled-components';
 
 //components
-import DetailCover from '../../components/Detail/DetailCover';
-import DetailTitle from '../../components/Detail/DetailTitle';
-import DetailInfo from '../../components/Detail/DetailInfo';
-import BoothMenu from '../../components/Detail/BoothMenu';
-import DetailReview from '../../components/Detail/DetailReview';
+import DetailCover from '../components/Detail/DetailCover';
+import DetailTitle from '../components/Detail/DetailTitle';
+import DetailInfo from '../components/Detail/DetailInfo';
+import BoothMenu from '../components/Detail/BoothMenu';
+import DetailReview from '../components/Detail/DetailReview';
 
 const BoothDetailPage = () => {
   return (
