+++ src/pages/PerfDetailPage.js
@@ -1,10 +1,10 @@
 import styled from 'styled-components';
 
 //components
-import DetailCover from '../../components/Detail/DetailCover';
-import DetailTitle from '../../components/Detail/DetailTitle';
-import DetailInfo from '../../components/Detail/DetailInfo';
-import DetailReview from '../../components/Detail/DetailReview';
+import DetailCover from '../components/Detail/DetailCover';
+import DetailTitle from '../components/Detail/DetailTitle';
+import DetailInfo from '../components/Detail/DetailInfo';
+import DetailReview from '../components/Detail/DetailReview';
 
 const PerfDetailPage = () => {
   return (
